### 2019-08-10 #1638 @whedon
[PRE REVIEW]: Welcome to the tidyverse
**Submitting author:** @hadley (<a href="http://orcid.org/0000-0003-4757-117X">Hadley Wickham</a>)
**Repository:** <a href="http://github.com/tidyverse/tidyverse" target ="_blank">http://github.com/tidyverse/tidyverse</a>
**Version:** v1.3.0
