### 2019-08-10 #1638 @arfon
:wave: @hadley - thanks for your submission to JOSS. The editor you suggested to handle this submission (@karthik) is currently on leave from JOSS for a couple of months so I'm going to ask one of our other editors to handle this submission.

@mbobra - would you be willing to handle this submission?
