### 2019-08-28 #1686 @whedon
[REVIEW]: Welcome to the tidyverse
**Submitting author:** @hadley (<a href="http://orcid.org/0000-0003-4757-117X">Hadley Wickham</a>)
**Repository:** <a href="http://github.com/tidyverse/tidyverse" target ="_blank">http://github.com/tidyverse/tidyverse</a>
**Version:** v1.3.0

**If I understand correctly, my role is to perform these checks for each of the twenty tidyverse packages and also the _tidyverse R_ package. Therefore I have provided information for each item on the checklist below for each package. Please let me know if I have misunderstood anything.**

  - [x] tidyverse
    - [x] broom
    - [x] dplyr
    - [x] feather
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr
    - [x] jsonlite
    - [x] lubridate
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse
    - [ ] xml2
    - [x] tidyverse

    - [x] broom
    - [x] dplyr
    - [x] feather
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr
    - [x] jsonlite
    - [x] lubridate
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse
    - [x] xml2
    - [x] broom
    - [x] dplyr
    - [x] feather
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr
    - [x] jsonlite
    - [x] lubridate
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse
    - [x] xml2
    - [x] broom
    - [x] dplyr
    - [x] feather
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr
    - [x] jsonlite
    - [x] lubridate
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse
    - [x] xml2

    - [x] broom
    - [x] dplyr
    - [x] feather
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr
    - [x] jsonlite
    - [x] lubridate
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse
    - [x] xml2
    - [x] broom
    - [x] dplyr
    - [x] feather
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr
    - [x] jsonlite
    - [x] lubridate
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse
    - [x] xml2
    - [x] broom
    - [x] dplyr
    - [ ] feather (documentation for `write_feather`---aliased as the documentation for `read_feather`---is missing an example that uses the `write_feather` function)
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr
    - [x] jsonlite
    - [x] lubridate
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse
    - [x] xml2
    - [x] broom
    - [x] dplyr
    - [x] feather
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr
    - [x] jsonlite
    - [x] lubridate
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse
    - [x] xml2
    - **Although all packages contain automated tests, some of the packages do not appear to have relatively high coverage (i.e. greater than 80%). This might be due to issues due with using continuous integration services (e.g. limited run time, disk space, or bandwidth). I have listed below some functions in tidyverse packages that might benefit from unit tests in case this is helpful.**
    - [x] broom
    - [x] dplyr
    - [x] feather (coverage not reported but contains tests for main functions)
    - [x] forcats
    - [x] ggplot2
    - [x] haven
    - [x] hms
    - [x] httr ([48% coverage coverage](https://codecov.io/gh/r-lib/httr?branch=master)): it might be useful to add tests for the functions: `cache_info`, `rerequest`, `httr_options`, `curl_docs`, `set_config`, `reset_config`, `with_config`, `cookies`, `set_cookies`, `httr_dr`, `set_envvar`, `get_envvar`, `BROWSE`, `oauth_endpoints` (endpoints for linkedin, twitter, vimeo, facebook, github, and azure), `progress`, `use_proxy`, `http_type`, `user_agent`, `verbose`, `write_disk`, `write_memory`, `write_stream`.
    - [x] jsonlite ([63% coverage coverage](https://codecov.io/github/jeroen/jsonlite?branch=master)): it might be useful to add tests for the functions: `stream_in`, `stream_out` (also verifies internal `apply_by_pages` and `stream_out_page` functions), `serializeJSON` (also verifies `pack` and `fixNativeSymbol`), `flatten`, `loadpkg`, `rbind_pages`, `unbox`.
    - [x] lubridate ([72% coverage coverage](https://codecov.io/gh/tidyverse/lubridate)): it might be useful to add tests for the functions: `fit_to_timeline`, `pretty_dates` (with seconds, minutes, hours, and months for coverage of internal `pretty_sec`, `pretty_min`, `pretty_hour`, `pretty_month` functions), `semester`, `dst`, `Date`.
    - [x] modelr
    - [x] purr
    - [x] readr
    - [x] rvest ([43% coverage](https://codecov.io/gh/tidyverse/rvest?branch=master)): it might useful to add tests for functions for forms (`set_values`, `submit_form`, `google_form`), sessions (`html_session`, `jump_to`, `follow_link`, `session_history`, `back`, `*.session` methods), and encoding (`guess_encoding`, `repair_encoding`).
    - [x] readxl
    - [x] stringr
    - [x] tibble
    - [x] tidyr
    - [x] tidyverse [58% coverage](https://codecov.io/gh/tidyverse/tidyverse?branch=master): perhaps it might be useful to add tests for: `tidyverse_deps`, `tidyverse_logo`.
    - [x] xml2
    **I have checked the CODE_OF_CONDUCT, DESCRIPTION, README, .github/*.md and vignette files for these three pieces of information, and listed below which information might be missing for each tidyverse package. The left check box corresponds to (1), the middle check box to (2) and the third check box to (3). Please let me know if I have missed anything.**
    - [x] [x] [x] broom
    - [x] [x] [x] dplyr
    - [ ] [x] [ ] feather
    - [x] [x] [x] forcats
    - [x] [x] [x] ggplot2
    - [x] [x] [x] haven
    - [x] [x] [x] hms
    - [ ] [x] [ ] httr
    - [ ] [x] [ ] jsonlite
    - [x] [x] [x] lubridate
    - [x] [x] [x] modelr
    - [x] [x] [x] purr
    - [x] [x] [x] readr
    - [x] [x] [x] rvest
    - [x] [x] [x] readxl
    - [x] [x] [x] stringr (information available via files in tidyverse/.github when opening new issue)
    - [x] [x] [x] tibble
    - [x] [x] [x] tidyr
    - [x] [x] [x] tidyverse
    - [ ] [x] [ ] xml2
