### 2019-09-06 #1686 @jeffreyhanson
@hadley, I'm very sorry if this is a silly question and I don't want to waste your time, but I just wanted to double check if the pdf version of the manuscript linked above has the correct version of your manuscript? I was noticed that the `paper.Rmd` file has been updated more recently than the `paper.md` file in the repository which (if I understand correctly) is being used to generate the pdf version of the manuscript (e.g. line 117 of the documents are different).
