### 2019-09-06 861c61f
* Render paper with `rmarkdown::render('~/tidyverse/vignettes/paper.Rmd')`
