+++ vignettes/paper.md
@@ -74,7 +74,7 @@ authors:
 affiliations:
  - name: RStudio
    index: 1
- - name: datatools
+ - name: cynkra
    index: 2
  - name: Redbubble
    index: 3
@@ -104,22 +104,22 @@ bibliography: paper.bib
 
 ![](tidyverse-logo.png)
 
-At a high level, the tidyverse is a language for solving data science challenges with R code. Its primary goal is to facilitate a conversation between a human and a computer about data. Less abstractly, the tidyverse is a collection of R packages that share an underlying design philosophy, grammar, and data structures, with the aim that learning one package makes it easier to learn the next. 
+At a high level, the tidyverse is a language for solving data science challenges with R code. Its primary goal is to facilitate a conversation between a human and a computer about data. Less abstractly, the tidyverse is a collection of R packages that share a high-level design philosophy and low-level grammar and data structures, so that learning one package should make it easier to learn the next. 
 
-The tidyverse encompasses the repeated tasks at the heart of every data science project: data import, tidying, manipulation, visualisation, and programming. We expect that almost every project will use multiple domain-specific packages outside of the tidyverse: our goal is to provide tooling for the most common challenges, not to solve every possible problem. Notably, the tidyverse doesn't include tools for statistical modelling or communication. These toolkits are critical for data science, but big enough domains that they merit separate treatment.
+The tidyverse encompasses the repeated tasks at the heart of every data science project: data import, tidying, manipulation, visualisation, and programming. We expect that almost every project will use multiple domain-specific packages outside of the tidyverse: our goal is to provide tooling for the most common challenges, not to solve every possible problem. Notably, the tidyverse doesn't include tools for statistical modelling or communication. These toolkits are critical for data science, but are so large that they merit separate treatment.
 
 This paper describes the tidyverse package, the components of the tidyverse, and some of the underlying philosophy.
 
 # Tidyverse package
 
-As well as being a collection of packages, the tidyverse itself is also a package. This provides a convenient way of downloading all tidyverse packages with a single R command:
+The tidyverse is a collection of packages that can easily be installed with a single "meta"-package also called tidyverse. This provides a convenient way of downloading all tidyverse packages with a single R command:
 
 
 ```r
 install.packages("tidyverse")
 ```
 
-The core tidyverse includes the packages that you're likely to use in everyday data analyses, and these are loaded when you load the tidyverse:
+The core tidyverse includes the packages that you're likely to use in everyday data analyses, and these are attached when you attach the tidyverse package:
 
 
 ```r
@@ -134,7 +134,9 @@ library(tidyverse)
 #> ✖ dplyr::lag()    masks stats::lag()
 ```
 
-As of tidyverse 1.2.0, the core packages include ggplot2 [@ggplot2], dplyr [@R-dplyr], tidyr [@R-tidyr], readr [@R-readr], purrr [@R-purrr], tibble [@R-tibble], stringr [@R-stringr], and forcats [@R-forcats]. Non-core packages are installed with `install.packages("tidyverse")`, but are not loaded with `library(tidyverse)`. They play more specialised roles, so are loaded individually as needed.
+As of tidyverse 1.2.0, the core packages include ggplot2 [@ggplot2], dplyr [@R-dplyr], tidyr [@R-tidyr], readr [@R-readr], purrr [@R-purrr], tibble [@R-tibble], stringr [@R-stringr], and forcats [@R-forcats]. 
+
+Non-core packages are installed with `install.packages("tidyverse")`, but are not attached by `library(tidyverse)`. They play more specialised roles, so will be individually loaded by the analyst as needed.
 
 # Components
 
@@ -142,23 +144,26 @@ How do the component packages of the tidyverse fit together? We use the model of
 
 ![](data-science.png)
 
[…]
