### 2019-09-06 9a4cbe9
Update paper.md
