### 2019-09-06 #1686 @hadley
Rendered preview is definitely ok now — sorry for the mess up.
