### 2019-09-11 #1686 @jeffreyhanson
The manuscript "Welcome to the tidyverse" provides an excellent overview of the tidyverse---an ecosystem of _R_ packages that share a vision for data science---and the _tidyverse R_ package---an _R_ package that helps manage this ecosystem---that will serve as a useful introduction for new data scientists. Currently, the tidyverse collection of packages spans eight core packages and twelve non-core packages. The core packages provide tools for importing (_readr R_ package), organizing (_tibble R_ package), manipulating (_dplyr_, _forcats_, _purr_, _stringr_, _tidyr R_ packages), and visualizing  data (_ggplot2 R_ package). The non-core packages provide more specialized functionalities, such as importing and exporting data from specific formats (e.g. _feather R_ package), manipulating specific data types (e.g. _lubridate R_ package), and converting complex and verbose model outputs into simple tabular formats (e.g. _broom R_ package). In my view, the manuscript is well-written for a broad audience. The tidyverse has had a profound influence over data science, and I look forward to seeing where it takes us in the future.

I have organized this review into three sections: **Major issues**, **Minor issues**, and **Assorted comments**. To be clear, the issues listed under **Assorted comments** are incredibly minor and do not substantially affect the quality of the manuscript. As a consequence, they can be ignored if the authors do not think changes are needed. Since the pdf version of the manuscript does not contain line numbers, I will refer to the line numbers as they appear in the [source rmarkdown document at the time of writing this review](https://github.com/tidyverse/tidyverse/blob/0271bf7f0a65870167fb3a897c4daa97905b31ba/vignettes/paper.Rmd).

@karthik, I have a few questions about exactly how some of the criteria in the **Reviewer checklist** should be applied to this particular submission (see the bold text in the checklist above). Once you've answered those questions, I can begin creating issues on the respective repositories if that's ok?

### Major issues

- Lines 115--131: In my opinion, the _tidyverse R_ package provides two additional functionalities that are not discussed in the manuscript. Firstly, attaching the _tidyverse R_ package will ensure that tidyverse packages are attached in the correct order. This is not trivial for beginners (e.g. attaching the _plyr_ and _dplyr R_ packages in the incorrect order will cause issues). Secondly, attaching the _tidyverse R_ package will inform the user about namespace conflicts in a concise and informative manner. This is very useful when many tidyverse and non-tidyverse packages are attached. For instance, conflicts arise when the _tidyverse R_ package is attached following the _raster R_ package because both the _tidyr R_ package---a core tidyverse package---and the _raster R_ package contain `extract()` functions. I think the _tidyverse R_ package is exceptionally useful when evaluated on its standalone merits---that is, disregarding the usefulness of the other packages that comprise the tidyverse---and I think readers would benefit from knowing about these functionalities.

### Minor issues

- Line 109: Perhaps it is worth providing a citation for R? I realize that many readers will likely already be familiar with R to some extent, but perhaps this be useful for readers that are less familiar with R?
- Lines 109--113: This manuscript is about describing (i) the collection of packages referred to as the tidyverse (covered in the **Components** and **Philosophy** sections) and (ii) the _tidyverse R_ package (covered in the **Tidyverse package** section). The **Summary** section does a fantastic job of summarizing the tidyverse collection of packages but I feel like it could provide more information about the _tidyverse R_ package itself. Perhaps the purpose of the _tidyverse R_ package could be mentioned in the **Summary** section?
- Lines 123--129: Is the _tidyverse R_ package not a core tidyverse package? It is not listed as a core package in Line 129. Maybe it would be useful to explicitly state if the _tidyverse R_ package is a core package or not?
- Line 131: I wonder if it is worth listing all of the non-core packages in the tidyverse? I suggest this for two reasons. Firstly, it is noted that one of the main purposes for the _tidyverse R_ package is that it provides a convenient method to install all of the packages in the tidyverse. Since there over ten non-core packages listed on the tidyverse package's REAMDE file, I suspect that listing all of these packages here would help the reader appreciate this functionality. Secondly, if one of the main aims of this manuscript is to describe the tidyverse, then I think it would be useful to list all of the packages that comprise the tidyverse. Although additional packages may be added to the tidyverse in the future, I think this would helpful to the reader.
- Throughout manuscript: The **Reviewer Checklist** requires that "the authors describe how this software compares to other commonly-used packages". This manuscript does not make such comparisons. For instance, the tidyverse _readr R_ package provides functions for importing data similar to the _utils_ (`read.table()`) and _data.table R_ packages. If this manuscript were specifically about the _readr R_ package then I would expect such comparisons to be present, but I personally think that detailed comparisons for all tidyverse packages would distract from the main purpose of the manuscript. I raise this issue here because it is on the checklist.

### Assorted comments

- Line 111: ", not to solve every possible problem". Perhaps a semicolon or em-dash would be more appropriate here than a comma?
- Line 117: Perhaps "downloading and installing" might be more appropriate than "downloading", since the following _R_ code uses `install.packages`?
- Line 129: Perhaps "tidyverse version 1.2.0" to make it clear that this sentence is referring to a specific version of the _tidyverse R_ package?
- Line 129: Is the ordering of these packages significant? If not, perhaps these packages could be listed alphabetically?
- Line 139: Perhaps "application programming interface (API)" instead of "API" for clarity?
- Line 139: This is a bit nitpicky but I do not think that data are necessarily "stored" in an API? I would say that data are accessed through an API? Apologies if I'm just plain wrong.
- Line 139: My impression is that terms like "data frame" and "flat file" will be very useful for individuals with experience in programming _R_, _Python_, or _SQL_. If such individuals are the intended audience for this manuscript, then using such terminology is ideal. If a broader audience is intended, then perhaps using different terms (or clarifying these terms) will be useful (e.g. terms like "tabular" or "spreadsheet" may be more informative for a broader audience).
- Lines 139 and 141: It appears that the term "data" is being used to refer to a singular entity, though some would argue that the term "data" is a plural. I do not know the grammar rules to which this journal subscribes, so I thought I would raise this point just in case.
- Line 139: Perhaps it might be useful to provide the full name of the format that corresponds to these file extensions (e.g. tab-separated values for tsv)?
- Line 139: I don't want to wade into the debate regarding if data is a singular or plural term, but I think that it might be useful to be consistent throughout the manuscript. My interpretation of the sentence containing the phrase "[...] import data stored in other common formats or retrieve it directly [...]" uses the term "data" to refer to both a plural and a singular entity. Perhaps the sentence might be clearer if written as something like "[...] import data from other common formats or the web directly"?
- Line 139 and elsewhere: The manuscript contains many embedded links which mean that URLs are not shown in the text (e.g. [like this](https://google.com)). I am not familiar enough with the correct reference style for this journal, so I thought I would raise this point just in case.
- Line 141: Perhaps "core tidyr package provides" might be useful to be consistent with the rest of the manuscript (e.g. "the core readr package", line 139)?
- References: The reference for Bache & Wickham (2014) has a lower case "r" in the title when referring to the R programming language.
