### 2019-09-23 6271023
Add paper author affiliation (#198)
* Change author affiliation.

* Fix actual affiliation
