### 2019-09-23 f746bbb
Add citation to R language.
