+++ vignettes/paper.bib
@@ -229,3 +229,12 @@
     note = {R package version 1.2.1},
     url = {https://CRAN.R-project.org/package=tidyverse},
   }
+
+@Manual{R-core,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2019},
+    url = {https://www.R-project.org/},
+  }
