### 2019-10-16 #1686 @karthik
> @karthik, I have a few questions about exactly how some of the criteria in the Reviewer checklist should be applied to this particular submission (see the bold text in the checklist above). Once you've answered those questions, I can begin creating issues on the respective repositories if that's ok?

Hi @jeffreyhanson Thanks so much for this exhaustive checkbox list. Please go ahead with opening issues in respective repos. I realize this is a huge undertaking as this is a large suite of packages.
