+++ vignettes/paper.Rmd
@@ -111,13 +111,13 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 
 At a high level, the tidyverse is a language for solving data science challenges with R code [@R-core]. Its primary goal is to facilitate a conversation between a human and a computer about data. Less abstractly, the tidyverse is a collection of R packages that share a high-level design philosophy and low-level grammar and data structures, so that learning one package makes it easier to learn the next. 
 
-The tidyverse encompasses the repeated tasks at the heart of every data science project: data import, tidying, manipulation, visualisation, and programming. We expect that almost every project will use multiple domain-specific packages outside of the tidyverse: our goal is to provide tooling for the most common challenges, not to solve every possible problem. Notably, the tidyverse doesn't include tools for statistical modelling or communication. These toolkits are critical for data science, but are so large that they merit separate treatment.
+The tidyverse encompasses the repeated tasks at the heart of every data science project: data import, tidying, manipulation, visualisation, and programming. We expect that almost every project will use multiple domain-specific packages outside of the tidyverse: our goal is to provide tooling for the most common challenges; not to solve every possible problem. Notably, the tidyverse doesn't include tools for statistical modelling or communication. These toolkits are critical for data science, but are so large that they merit separate treatment. The tidyverse package allows users to install all tidyverse packages with a single command.
 
 This paper describes the tidyverse package, the components of the tidyverse, and some of the underlying philosophy.
 
 # Tidyverse package
 
-The tidyverse is a collection of packages that can easily be installed with a single "meta"-package, which called "tidyverse". This provides a convenient way of downloading all tidyverse packages with a single R command:
+The tidyverse is a collection of packages that can easily be installed with a single "meta"-package, which called "tidyverse". This provides a convenient way of downloading and installing all tidyverse packages with a single R command:
 
 ```{r, eval = FALSE}
 install.packages("tidyverse")
@@ -129,9 +129,16 @@ The core tidyverse includes the packages that you're likely to use in everyday d
 library(tidyverse)
 ```
 
-As of tidyverse 1.2.0, the core packages include ggplot2 [@ggplot2], dplyr [@R-dplyr], tidyr [@R-tidyr], readr [@R-readr], purrr [@R-purrr], tibble [@R-tibble], stringr [@R-stringr], and forcats [@R-forcats]. 
+As of tidyverse version 1.2.0, the core packages include dplyr [@R-dplyr], forcats [@R-forcats], ggplot2 [@ggplot2], purrr [@R-purrr], readr [@R-readr], stringr [@R-stringr], tibble [@R-tibble], and tidyr [@R-tidyr]. 
+<!--  JH: Firstly, attaching the tidyverse R package will ensure that tidyverse packages are attached in the correct order. This is not trivial for beginners (e.g. attaching the plyr and dplyr R packages in the incorrect order will cause issues). Secondly, attaching the tidyverse R package will inform the user about namespace conflicts in a concise and informative manner. This is very useful when many tidyverse and non-tidyverse packages are attached. -->
 
-Non-core packages are installed with `install.packages("tidyverse")`, but are not attached by `library(tidyverse)`. They play more specialised roles, so will be attached by the analyst as needed.
+<!-- Is the tidyverse R package not a core tidyverse package? It is not listed as a core package in Line 129. Maybe it would be useful to explicitly state if the tidyverse R package is a core package or not? --> 
+<!-- MGA re. above: This would make the tidyverse like recursively core...clearly we're not describing it well, but I'm not sure how to clarify -->
+
[…]
