### 2019-10-28 #1686 @karthik
@batpigandme No rush, but when possible can you give me a rough timeline for when I should check in with you on revisions?
