### 2019-10-29 #1686 @batpigandme
Hi @karthik,
We have team work week this week, but it's high on our priority list, and we're aiming to finish up next week.

Do you happen to have a particularly good example of what responding to feedback should look like? Do we need to resolve all the checkboxes.

Thanks for your help, and your patience!
