### 2019-10-29 #1686 @karthik
@jeffreyhanson I'm answering a few more questions here.

> This repository contains the source code for the tidyverse R package. It does not contain the source code for all the tidyverse packages. However, the README file in this repository contains links to other GitHub repositories which contain the tidyverse packages. Personally, I believe this approach is more sensible than using a single repository for all tidyverse packages, but this particular approach is technically not acceptable based on my (super literal) interpretation of this criteria. @karthik, how should I proceed?

No need to interpret this super literally. As long as individual packages contain source code for each package, we are good to check this off.

> Although not all tidyverse packages contain a LICENSE file, it is worth mentioning that all tidyverse packages contain a DESCRIPTION file which indicates the software license by name. @karthik, do all packages still require a LICENSE file?

Don't all R packages require a LICENSE file? http://r-pkgs.had.co.nz/description.html#license even when the license itself is mentioned in the README. Can you point me to ones that don't?

> I have checked whether the submitting author is listed as an author in the DESCRIPTION file of each tidyverse package. I did not differentiate based on the role (e.g. aut, cre, ctb). @karthik, since this particular submission pertains to a collection of packages is it acceptable if @hadley is not an author of every single package?

That's fine in this case. It's clear that the Tidyverse idea itself is from Hadley, and this collection represents core packages that are part of that philosophy. So in this case I'd say it's perfectly fine that Hadley is not listed as an author on all of the packages.
