### 2019-10-29 #1686 @karthik
@batpigandme  Here are a couple of good examples of responses:

https://github.com/openjournals/joss-reviews/issues/964
https://github.com/openjournals/joss-reviews/issues/741

> Do we need to resolve all the checkboxes.

Yes. or provide a reasonable explanation for ones that cannot be resolved so that the reviews can respond.
