### 2019-10-30 #1686 @hadley
@karthik (just responding quickly to license question): From R-exts:

> Whereas you should feel free to include a license file in your source distribution, please do not arrange to install yet another copy of the GNU COPYING or COPYING.LIB files but refer to the copies on https://www.R-project.org/Licenses/ and included in the R distribution (in directory share/licenses). Since files named LICENSE or LICENCE will be installed, do not use these names for standard license files. To include comments about the licensing rather than the body of a license, use a file named something like LICENSE.note.

We work around this in (e.g.) `use_gpl_license()` by including a `.Rbuildignore`d `LICENSE.md` which R CMD check don't see locally and isn't included in the package bundle (and still allows you to use `LICENSE` for the (e.g.) template fields needed for to make the MIT license compatible with CRAN's requirements.)

We include license files in new packages, but we haven't made a systematic effort to include them in all older packages, so this is a good prompt for us to do so.
