### 2019-11-06 d91cbcb
Commit review comments from ldecicco-USGS
