### 2019-11-06 26f822d
Respond to first review comment
