+++ vignettes/paper.Rmd
@@ -129,7 +129,7 @@ The core tidyverse includes the packages that you're likely to use in everyday d
 library(tidyverse)
 ```
 
-As of tidyverse version 1.2.0, the core packages include dplyr [@R-dplyr], forcats [@R-forcats], ggplot2 [@ggplot2], purrr [@R-purrr], readr [@R-readr], stringr [@R-stringr], tibble [@R-tibble], and tidyr [@R-tidyr]. Using `library(tidyverse)` ensures that the packages are attached in the correct order, and provides a concise summary of namespace conflicts. 
+This is a convenient shortcut for attaching the cores packages, produces a short report that gives the package versions you're using, and succintly informs you of any conflicts with packages you've previously loaded. As of tidyverse version 1.2.0, the core packages include dplyr [@R-dplyr], forcats [@R-forcats], ggplot2 [@ggplot2], purrr [@R-purrr], readr [@R-readr], stringr [@R-stringr], tibble [@R-tibble], and tidyr [@R-tidyr]. 
 
 <!--  JH: Firstly, attaching the tidyverse R package will ensure that tidyverse packages are attached in the correct order. This is not trivial for beginners (e.g. attaching the plyr and dplyr R packages in the incorrect order will cause issues). Secondly, attaching the tidyverse R package will inform the user about namespace conflicts in a concise and informative manner. This is very useful when many tidyverse and non-tidyverse packages are attached. -->
 <!-- MGA: I want to take care of the above with LD comment below -->
@@ -143,7 +143,7 @@ As of tidyverse version 1.2.0, the core packages include dplyr [@R-dplyr], forca
 
 Non-core packages are installed with `install.packages("tidyverse")`, but are not attached by `library(tidyverse)`. They play more specialised roles, so will be attached by the analyst as needed. The non-core packages are: blob [@R-blob], glue [@R-glue], googledrive [@R-googledrive], haven [@R-haven], hms [@R-hms], lubridate [@R-lubridate], magrittr [@R-magrittr], modelr [@R-modelr], readxl [@R-readxl], reprex [@R-reprex], and rvest [@R-rvest].
 
-The tidyverse package is designed for interactive use, and should not be imported for package-development purposes. The `tidyverse_deps()` function returns a tibble of the package's dependencies. Its one argument, `recursive`, can be set to `TRUE` to include the dependencies of the tidyverse packages--totalling 87, as of this writing. We recommend that package developers import, suggest, or depend on component packages.
+The tidyverse package is designed with an eye for teaching: `install.packages("tidyverse")` gets you a "batteries-included" set of 87 packages (at time of writing). This large set of dependencies means that it is not appropriate to use the tidyverse package within another package; instead we recommend that package authors import only the specific packages that they use.
 
 # Components
 
