+++ vignettes/paper.bib
@@ -62,6 +62,14 @@
   url = {https://CRAN.R-project.org/package=dplyr}
 }
 
+@Manual{R-feather,
+  title = {feather: R Bindings to the Feather API},
+  author = {Hadley Wickham},
+  year = {2019},
+  note = {R package version 0.3.3},
+  url = {https://CRAN.R-project.org/package=feather},
+}
+
 @Manual{R-forcats,
   title = {{forcats}: Tools for Working with Categorical Variables (Factors)},
   author = {Hadley Wickham},
