### 2019-11-07 cc9ebba
Add tidy github and LICENSE.md
