+++ vignettes/paper.bib
@@ -119,7 +119,7 @@
 }
 
 @Manual{R-magrittr,
-  title = {{magrittr}: A Forward-Pipe Operator for R},
+  title = {{magrittr}: A Forward-Pipe Operator for {R}},
   author = {Stefan Milton Bache and Hadley Wickham},
   year = {2014},
   note = {R package version 1.5},
