### 2019-11-08 11a8116
Minor copy edits, and removal of comments.
