+++ vignettes/paper.bib
@@ -246,3 +246,37 @@
     year = {2019},
     url = {https://www.R-project.org/},
   }
+
+@Manual{SAS-Soft,
+    author = {{SAS} Institute Inc.},
+    title = {{SAS/STAT} Software, Version~9.3},
+    year = 2010,
+    organization = {{SAS} Institute Inc.},
+    address = {Cary, NC},
+    url = {http://www.sas.com/},
+  }
+
+@Manual{SPSS-Soft,
+    author={{SPSS} Inc.},
+    title={{IBM} {SPSS} Statistics~21},
+    year=2012,
+    organization={SPSS Inc.},
+    address={Chicago, IL},
+    url={http://www.spss.com/},
+  }
+
+@Manual{Stata-Soft,
+    author = {StataCorp.},
+    title = {{Stata} Statistical Software: Release 16},
+    address = {College Station, TX},
+    publisher = {StataCorp LLC},
+    year = {2019},
+  }
+
+@Manual{R-data-table,
+    title = {{data.table}: Extension of `data.frame`},
+    author = {Matt Dowle and Arun Srinivasan},
+    year = {2019},
+    note = {R package version 1.12.6},
+    url = {https://CRAN.R-project.org/package=data.table},
+  }
