### 2019-11-08 ddd8839
Add comparison packages and references.
