+++ vignettes/paper.bib
@@ -247,33 +247,35 @@
     url = {https://www.R-project.org/},
   }
 
-@Manual{SAS-Soft,
-    author = {{SAS} Institute Inc.},
-    title = {{SAS/STAT} Software, Version~9.3},
-    year = 2010,
-    organization = {{SAS} Institute Inc.},
-    address = {Cary, NC},
-    url = {http://www.sas.com/},
+@article{Bioconductor,
+    author = {Gentleman, Robert C and Carey, Vincent J and Bates, Douglas M and Bolstad, Ben and Dettling, Marcel and Dudoit, Sandrine and Ellis, Byron and Gautier, Laurent and Ge, Yongchao and Gentry, Jeff and others},
+    year = {2004},
+    title = {Bioconductor: open software development for computational biology and bioinformatics},
+    journal = {Genome Biology},
+    volume = {5},
+    number = {10},
+    pages = {R80},
+}
+
+@article{Huber2015,
+    author = {W. Huber and V. J. Carey and R. Gentleman and S. Anders and M. Carlson and B. S. Carvalho and H. C. Bravo and S. Davis and L. Gatto and T. Girke and R. Gottardo and F. Hahne and K. D. Hansen and R. A. Irizarry and M. Lawrence and M. I. Love and J. MacDonald and V. Obenchain and A. K. {Ole's} and H. {Pag`es} and A. Reyes and P. Shannon and G. K. Smyth and D. Tenenbaum and L. Waldron and M. Morgan},
+    title = {{O}rchestrating high-throughput genomic analysis with {B}ioconductor},
+    journal = {Nature Methods},
+    year = {2015},
+    volume = {12},
+    number = {2},
+    pages = {115--121},
+    url = {http://www.nature.com/nmeth/journal/v12/n2/full/nmeth.3252.html},
   }
 
-@Manual{SPSS-Soft,
-    author={{SPSS} Inc.},
-    title={{IBM} {SPSS} Statistics~21},
-    year=2012,
-    organization={SPSS Inc.},
-    address={Chicago, IL},
-    url={http://www.spss.com/},
-  }
+@Manual{R-Biobase,
+  title = {Biobase: Base functions for Bioconductor},
+  author = {R. Gentleman and V. Carey and M. Morgan and S. Falcon},
+  year = {2019},
+  note = {R package version 2.46.3},
+}
 
-@Manual{Stata-Soft,
-    author = {StataCorp.},
-    title = {{Stata} Statistical Software: Release 16},
-    address = {College Station, TX},
-    publisher = {StataCorp LLC},
-    year = {2019},
-  }
-
-@Manual{R-data-table,
+@Manual{R-datatable,
     title = {{data.table}: Extension of `data.frame`},
     author = {Matt Dowle and Arun Srinivasan},
     year = {2019},
