+++ vignettes/paper.md
@@ -97,7 +97,7 @@ affiliations:
    index: 11
  - name: University of Wisconsin, Madison
    index: 12
-date: 17 May 2019
+date: 08 November 2019
 bibliography: paper.bib
 ---
 
