### 2019-11-08 793745c
Reference tweaks
