+++ vignettes/paper.bib
@@ -57,13 +57,13 @@
 @Manual{R-dplyr,
   title = {{dplyr}: A Grammar of Data Manipulation},
   author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
-  year = {2018},
-  note = {R package version 0.7.8},
+  year = {2019},
+  note = {R package version 0.8.3},
   url = {https://CRAN.R-project.org/package=dplyr}
 }
 
 @Manual{R-feather,
-  title = {feather: R Bindings to the Feather API},
+  title = {feather: {R} Bindings to the Feather {API}},
   author = {Hadley Wickham},
   year = {2019},
   note = {R package version 0.3.3},
@@ -73,16 +73,16 @@
 @Manual{R-forcats,
   title = {{forcats}: Tools for Working with Categorical Variables (Factors)},
   author = {Hadley Wickham},
-  year = {2018},
-  note = {R package version 0.3.0},
+  year = {2019},
+  note = {R package version 0.4.0},
   url = {https://CRAN.R-project.org/package=forcats},
 }
 
 @Manual{R-ggplot2,
   title = {{ggplot2}: Create Elegant Data Visualisations Using the Grammar of Graphics},
   author = {Hadley Wickham and Winston Chang and Lionel Henry and Thomas Lin Pedersen and Kohske Takahashi and Claus Wilke and Kara Woo and Hiroaki Yutani},
-  note = {R package version 3.1.1},
-  year = {2018},
+  note = {R package version 3.2.1},
+  year = {2019},
   url = {https://CRAN.R-project.org/package=ggplot2}
 }
 
@@ -95,7 +95,7 @@
 }
 
 @Manual{R-haven,
-    title = {{haven}: Import and Export "SPSS", "Stata" and "SAS" Files},
+    title = {{haven}: Import and Export {SPSS}, {Stata}, and {SAS} Files},
     author = {Hadley Wickham and Evan Miller},
     year = {2018},
     note = {R package version 2.0.0},
@@ -137,8 +137,8 @@
 @Manual{R-purrr,
   title = {{purrr}: Functional Programming Tools},
   author = {Lionel Henry and Hadley Wickham},
-  year = {2018},
-  note = {R package version 0.2.5},
+  year = {2019},
+  note = {R package version 0.3.3},
   url = {https://CRAN.R-project.org/package=purrr},
 }
 
@@ -153,32 +153,32 @@
 @Manual{R-readxl,
   title = {{readxl}: Read Excel Files},
   author = {Hadley Wickham and Jennifer Bryan},
-  year = {2018},
-  note = {R package version 1.2.0},
+  year = {2019},
+  note = {R package version 1.3.1},
   url = {https://CRAN.R-project.org/package=readxl},
 }
 
 @Manual{R-reprex,
   title = {{reprex}: Prepare Reproducible Example Code via the Clipboard},
   author = {Jennifer Bryan and Jim Hester and David Robinson and Hadley Wickham},
-  year = {2018},
-  note = {R package version 0.2.1},
+  year = {2019},
+  note = {R package version 0.3.0},
   url = {https://CRAN.R-project.org/package=reprex}
 }
 
 @Manual{R-rvest,
   title = {{rvest}: Easily Harvest (Scrape) Web Pages},
   author = {Hadley Wickham},
-  year = {2016},
-  note = {R package version 0.3.2},
+  year = {2019},
+  note = {R package version 0.3.4},
   url = {https://CRAN.R-project.org/package=rvest},
 }
 
 @Manual{R-stringr,
   title = {{stringr}: Simple, Consistent Wrappers for Common String Operations},
   author = {Hadley Wickham},
-  year = {2018},
[…]
