### 2019-11-08 #1686 @batpigandme
Thank you so much, @jeffreyhanson and @ldecicco-USGS, for you thoughtful reviews, and, @karthik, for your guidance throughout this process.
Your input has made the paper much stronger.

We've responded to your feedback piece by piece in three markdown documents:

* @ldecicco-USGS's [here](https://github.com/tidyverse/tidyverse/blob/master/vignettes/review-ldecicco-USGS.md),
* @jeffreyhanson's [here](https://github.com/tidyverse/tidyverse/blob/master/vignettes/review-jeffreyhanson.md), and
* edits for the review checklist [here](https://github.com/tidyverse/tidyverse/blob/master/vignettes/review-checklist.md).

Of course, if you prefer, we're happy to include the bodies of these documents as comments in this issue.

Thank you again for your time and patience, and please let us know if there's anything you think we've overlooked, our outstanding items that we need to address.
