+++ .travis.yml
@@ -11,6 +11,7 @@ matrix:
     - Rscript -e 'covr::codecov()'
     before_cache:
     - Rscript -e 'remotes::install_cran("pkgdown")'
+    - Rscript -e 'remotes::install_github("rstudio/rmarkdown")' # https://github.com/rstudio/rmarkdown/pull/1668
     - Rscript -e 'remotes::install_github("tidyverse/tidytemplate")'
     deploy:
       provider: script
