### 2019-11-11 #1686 @ldecicco-USGS
The responses look fantastic to me, and now I'm super excited to see the development of the Tidyverse design guide.... 👯 💯 🎉
