### 2019-11-11 #1686 @karthik
Thanks @batpigandme and @hadley!

> Of course, if you prefer, we're happy to include the bodies of these documents as comments in this issue.

Since it's part of the version history of Tidyverse, I think it's fine where it is as this review is tied to the repo.

I'll wait for @jeffreyhanson and @ldecicco-USGS to respond to your comments.
@ldecicco-USGS let me know when you are ready to sign off.
