### 2019-11-14 #1686 @jeffreyhanson
I am delighted to review the revised version of the "Welcome to the tidyverse" manuscript. In my view, provides a wonderful introduction to the tidyverse and the description of the _tidyverse R_ package. It is exceptionally well-written for a broad audience. The authors have addressed all of my issues with the previous version of the manuscript. I could not find any major or minor issues with this work, and have very few improvements to suggest for the manuscript. I re-checked the unresolved JOSS checklist items I flagged in my initial review, and almost all of them have been addressed. I have listed the few remaining unresolved checklist items below. I apologise if these checklist items have actually been addressed and I was just looking in the wrong files.

I have organized this review into three sections: **Major issues**, **Minor issues**, **Assorted comments**, and **Remaining checklist items**. To be clear, the issues listed under **Assorted comments** are incredibly minor and do not substantially affect the quality of the manuscript. As a consequence, they can be ignored if the authors do not think changes are useful. The issues listed under **Remaining checklist items** pertain to the JOSS reviewer checklist.

### Major issues

- None.

### Minor issues

- None.

### Assorted comments

- Line 190--191: Perhaps it might be worth describing the pipe as an operator when introducing it here? For example, instead of "... the pipe, `%>%`, ...", perhaps "... the pipe operator,`%>%`, ..." might be more informative? On the other hand, the term "operator" may be too technical for a broad-audience? I raise this just in case this is helpful.

### Remaining checklist items

- The _xml2 R_ package does not appear to contain a license file.
- The _feather R_ package does not provide an example of the `write_feather` function (see [Rd file here](https://github.com/wesm/feather/blob/master/R/man/read_feather.Rd)). Given that this function represents half of the functionality provided by this package, I think it would be important to provide an example of this function. However, such example code may not be possible given CRAN policies regarding `\donttest` and `\dontcheck` blocks in example code.
- The _feather_, _httr_, and the _jsonlite R_ packages appear to be missing information describing how other people can contribute to the software and/or where they can find help with using the software (e.g. in README or CODE_OF_CONDUCT files). I'm sorry if this issue was addressed and I'm just wrong about this. I note that the repositories for these package are not under the tidyverse GitHub organization, and so making such changes may not be trivial.
