### 2019-11-14 #1686 @jeffreyhanson
Sorry for not mentioning this in the previous post, I reviewed the revised version by looking at the Rmarkdown document in the tidyverse/tidyverse repository. As such, it might be worth using the JOSS automatic thingummywhatsit bot to make an updated pdf version of the manuscript to double check that the references appear correctly.
