### 2019-11-15 4e3af74
Describe pipe as "operator".
