### 2019-11-15 #1686 @batpigandme
^^ whoops, premature PDF generation. Sorry about that!
