### 2019-11-15 #1686 @batpigandme
Thanks again for your comments!

The xml2, httr, jsonlite, and feather packages have a somewhat murky relationship to the tidyverse. These packages are mostly installed because we expect them to be used pretty frequently but are owned by other organisations. We'll work to get finish the checklist for these packages (since it's good practice regardless of this article), but in the interests of unblocking this article, are we ok to publish without them?
