### 2019-11-15 #1686 @hadley
It looks like compilation failed because an author doesn't have an affiliation? Is this required? We asked all authors to supply their preferred affiliation, and this author preferred not to have one.
