### 2019-11-15 #1686 @arfon
> It looks like compilation failed because an author doesn't have an affiliation? Is this required? We asked all authors to supply their preferred affiliation, and this author preferred not to have one.

Unfortunately this is a limitation of our current template. You can solve this by creating an affiliation of 'None' and associating it with the author like this: https://github.com/tidyverse/tidyverse/pull/205
