### 2019-11-15 6b97e35
Add affilitation to source Rmd
