+++ vignettes/paper.bib
@@ -78,6 +78,14 @@
   url = {https://CRAN.R-project.org/package=forcats},
 }
 
+@Article{R-jsonlie,
+  title = {The jsonlite Package: A Practical and Consistent Mapping Between {JSON} Data and {R} Objects},
+  author = {Jeroen Ooms},
+  journal = {arXiv:1403.2805 [stat.CO]},
+  year = {2014},
+  url = {https://arxiv.org/abs/1403.2805},
+}
+
 @Manual{R-ggplot2,
   title = {{ggplot2}: Create Elegant Data Visualisations Using the Grammar of Graphics},
   author = {Hadley Wickham and Winston Chang and Lionel Henry and Thomas Lin Pedersen and Kohske Takahashi and Claus Wilke and Kara Woo and Hiroaki Yutani},
