+++ vignettes/paper.bib
@@ -290,3 +290,11 @@
     note = {R package version 1.12.6},
     url = {https://CRAN.R-project.org/package=data.table},
   }
+
+@Manual{R-xml2,
+  title = {xml2: Parse {XML}},
+  author = {Hadley Wickham and Jim Hester and Jeroen Ooms},
+  year = {2019},
+  note = {R package version 1.2.2},
+  url = {https://CRAN.R-project.org/package=xml2},
+}
