+++ vignettes/paper.bib
@@ -231,7 +231,7 @@
   }
 
 @Manual{R-rmarkdown,
-    title = {{rmarkdown}: Dynamic Documents for R},
+    title = {{rmarkdown}: Dynamic Documents for {R}},
     author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
     year = {2018},
     note = {R package version 1.12},
