### 2019-11-15 4a41ba3
More minor tweaks
