+++ vignettes/paper.bib
@@ -78,7 +78,7 @@
   url = {https://CRAN.R-project.org/package=forcats},
 }
 
-@Article{R-jsonlie,
+@Article{R-jsonlite,
   title = {The jsonlite Package: A Practical and Consistent Mapping Between {JSON} Data and {R} Objects},
   author = {Jeroen Ooms},
   journal = {arXiv:1403.2805 [stat.CO]},
