### 2019-11-15 df6402c
Don't use unicode in paper
