### 2019-11-15 4bbe586
Make console width narrower
