+++ vignettes/paper.Rmd
@@ -110,7 +110,7 @@ bibliography: paper.bib
 
 ```{r, include = FALSE}
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>", width = 68)
-options(width = 68, cli.unicode = FALSE)
+options(width = 68, cli.unicode = FALSE, cli.width = 68)
 ```
 
 # Summary
