+++ vignettes/paper.md
@@ -136,12 +136,12 @@ The core tidyverse includes the packages that you're likely to use in everyday d
 
 ```r
 library(tidyverse)
-#> -- Attaching packages ------------------------------------------------------------- tidyverse 1.2.1 --
+#> -- Attaching packages --------------------------- tidyverse 1.2.1 --
 #> v ggplot2 3.2.1     v purrr   0.3.3
 #> v tibble  2.1.3     v dplyr   0.8.3
 #> v tidyr   1.0.0     v stringr 1.4.0
 #> v readr   1.3.1     v forcats 0.4.0
-#> -- Conflicts ---------------------------------------------------------------- tidyverse_conflicts() --
+#> -- Conflicts ------------------------------ tidyverse_conflicts() --
 #> x dplyr::filter() masks stats::filter()
 #> x dplyr::lag()    masks stats::lag()
 ```
