### 2019-11-15 32a5333
Use high-resolution images
