### 2019-11-15 #1686 @hadley
Sorry for all the rebuilds — just trying to get the last details of the code and images perfect
