### 2019-11-15 #1686 @karthik
@hadley You can also use http://whedon.theoj.org/ to compile the paper
