### 2019-11-15 25a4576
Fix Jeroen's affiliation (#206)
