### 2019-11-17 #1686 @jeffreyhanson
Yeah, I don't think the remaining checklist items should hold up publication. I'll sign off now - absolutely awesome work!
