### 2019-11-18 #1686 @karthik
@hadley Can you please provide the latest version number for publication?
