### 2019-11-18 afbb04d
Adds DOI to Wilkinson's Grammar of Graphics
