### 2019-11-18 #1686 @hadley
@karthik what do people usually do with R packages? I'd like for the next release of tidyverse to an include a citation to this paper, but that introduces a chicken and egg problem.

I'll add the missing DOI momentarily.
