### 2019-11-18 29d62fd
More reference tweaks
* Add DOI for grammar of graphics
* fix capitalisation of xml2/XML
