+++ vignettes/paper.bib
@@ -35,6 +35,7 @@
  isbn = {0387245448},
  publisher = {Springer-Verlag},
  address = {Berlin, Heidelberg},
+ doi = {10.1007/0-387-28695-0}
 }
 
 @Manual{R-blob,
@@ -292,7 +293,7 @@
   }
 
 @Manual{R-xml2,
-  title = {xml2: Parse {XML}},
+  title = {{xml2}: Parse {XML}},
   author = {Hadley Wickham and Jim Hester and Jeroen Ooms},
   year = {2019},
   note = {R package version 1.2.2},
