### 2019-11-18 #1686 @karthik
I sent a PR with the missing DOI.

> what do people usually do with R packages? I'd like for the next release of tidyverse to an include a citation to this paper, but that introduces a chicken and egg problem.

We're aware of this chicken and egg problem. For now, once the reviewers sign off, you archive the reviewed version on Zenodo, we add that DOI to your JOSS paper and then publish it. You can then add the JOSS citation to Tidyverse immediately and push a new release to GitHub/CRAN. That way people will see this paper in `citation()`.

JOSS does not publish a new paper with every version change, so you can still use this paper as a canonical reference.

If you really want to add the citation _before_ you archive, it is:

```
Wickham et al., (2019). Welcome to the Tidyverse. Journal of Open Source Software, 4(43), 1686, https://doi.org/10.21105/joss.01686
```
