### 2019-11-18 9f66009
Fix tidyverse logo link (#195)
