### 2019-11-18 #1686 @hadley
Ok, in that case can we use 1.3.0 for the version number? There are [no significant changes](https://github.com/tidyverse/tidyverse/blob/master/NEWS.md#tidyverse-1219000) between 1.2.1 and 1.3.0 except that 1.3.0 will include a citation to this paper.

Can you tell what is wrong with my DOI at https://github.com/tidyverse/tidyverse/blob/master/vignettes/paper.bib#L38 ?
