### 2019-11-18 #1686 @arfon
> Can you tell what is wrong with my DOI at https://github.com/tidyverse/tidyverse/blob/master/vignettes/paper.bib#L38 ?

I wonder if Whedon is fussy about case? Can you try updating the BibTeX key to `doi` rather than `Doi`?
