+++ vignettes/paper.bib
@@ -35,7 +35,7 @@
  isbn = {0387245448},
  publisher = {Springer-Verlag},
  address = {Berlin, Heidelberg},
- Doi = {10.1007/0-387-28695-0},
+ doi = {10.1007/0-387-28695-0}
 }
 
 @Manual{R-blob,
