### 2019-11-18 5bc76ee
Don't include "tool" packages in dependencies
