+++ man/tidyverse-package.Rd
@@ -6,7 +6,7 @@
 \alias{tidyverse-package}
 \title{tidyverse: Easily Install and Load the 'Tidyverse'}
 \description{
-\if{html}{\figure{logo.png}{options: align='right'}}
+\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
 
 The 'tidyverse' is a set of packages that work in
     harmony because they share common data representations and 'API'
