### 2019-11-18 d526d3d
Implement tidyverse_sitrep()
Fixes #203
