+++ man/tidyverse_sitrep.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/update.R
+\name{tidyverse_sitrep}
+\alias{tidyverse_sitrep}
+\title{Get a situation report on the tidyverse}
+\usage{
+tidyverse_sitrep()
+}
+\description{
+This function gives a quick overview of the versions of R and RStudio as
+well as all tidyverse packages. It's primarily designed to help you get
+a quick idea of what's going on when you're helping someone else debug
+a problem.
+}
