+++ NAMESPACE
@@ -6,6 +6,7 @@ export(tidyverse_conflicts)
 export(tidyverse_deps)
 export(tidyverse_logo)
 export(tidyverse_packages)
+export(tidyverse_sitrep)
 export(tidyverse_update)
 importFrom(broom,tidy)
 importFrom(dbplyr,db_collect)
