### 2019-11-18 74b9c7d
Bump all dependency versions
