+++ DESCRIPTION
@@ -23,29 +23,29 @@ Imports:
     broom (>= 0.5.2),
     cli (>= 1.1.0),
     crayon (>= 1.3.4),
-    dbplyr (>= 1.4.0),
-    dplyr (>= 0.8.0.1),
+    dbplyr (>= 1.4.2),
+    dplyr (>= 0.8.3),
     forcats (>= 0.4.0),
-    ggplot2 (>= 3.1.1),
-    haven (>= 2.1.0),
-    hms (>= 0.4.2),
-    httr (>= 1.4.0),
+    ggplot2 (>= 3.2.1),
+    haven (>= 2.2.0),
+    hms (>= 0.5.2),
+    httr (>= 1.4.1),
     jsonlite (>= 1.6),
     lubridate (>= 1.7.4),
     magrittr (>= 1.5),
-    modelr (>= 0.1.4),
-    pillar (>= 1.3.1),
-    purrr (>= 0.3.2),
+    modelr (>= 0.1.5),
+    pillar (>= 1.4.2),
+    purrr (>= 0.3.3),
     readr (>= 1.3.1),
     readxl (>= 1.3.1),
-    reprex (>= 0.2.1),
-    rlang (>= 0.3.4),
+    reprex (>= 0.3.0),
+    rlang (>= 0.4.1),
     rstudioapi (>= 0.10),
-    rvest (>= 0.3.3),
+    rvest (>= 0.3.5),
     stringr (>= 1.4.0),
-    tibble (>= 2.1.1),
-    tidyr (>= 0.8.3),
-    xml2 (>= 1.2.0)
+    tibble (>= 2.1.3),
+    tidyr (>= 1.0.0),
+    xml2 (>= 1.2.2)
 Suggests: 
     covr,
     feather,
