+++ revdep/README.md
@@ -2,15 +2,15 @@
 
 |field    |value                        |
 |:--------|:----------------------------|
-|version  |R version 3.5.2 (2018-12-20) |
-|os       |macOS Mojave 10.14.3         |
+|version  |R version 3.6.0 (2019-04-26) |
+|os       |macOS Mojave 10.14.6         |
 |system   |x86_64, darwin15.6.0         |
 |ui       |RStudio                      |
 |language |(EN)                         |
 |collate  |en_US.UTF-8                  |
 |ctype    |en_US.UTF-8                  |
 |tz       |America/Chicago              |
-|date     |2019-05-15                   |
+|date     |2019-11-18                   |
 
 # Dependencies
 
@@ -19,95 +19,98 @@
 |tidyverse    |1.2.1    |1.2.1.9000 |*  |
 |askpass      |1.1      |1.1        |   |
 |assertthat   |0.2.1    |0.2.1      |   |
-|backports    |1.1.4    |1.1.4      |   |
+|backports    |1.1.5    |1.1.5      |   |
 |base64enc    |0.1-3    |0.1-3      |   |
 |BH           |1.69.0-1 |1.69.0-1   |   |
 |broom        |0.5.2    |0.5.2      |   |
-|callr        |3.2.0    |3.2.0      |   |
+|callr        |3.3.2    |3.3.2      |   |
 |cellranger   |1.1.0    |1.1.0      |   |
 |cli          |1.1.0    |1.1.0      |   |
-|clipr        |0.6.0    |0.6.0      |   |
+|clipr        |0.7.0    |0.7.0      |   |
+|colorspace   |1.4-1    |1.4-1      |   |
 |crayon       |1.3.4    |1.3.4      |   |
-|curl         |3.3      |3.3        |   |
+|curl         |4.2      |4.2        |   |
 |DBI          |1.0.0    |1.0.0      |   |
-|dbplyr       |1.4.0    |1.4.0      |   |
-|digest       |0.6.18   |0.6.18     |   |
-|dplyr        |0.8.0.1  |0.8.0.1    |   |
-|ellipsis     |0.1.0    |0.1.0      |   |
-|evaluate     |0.13     |0.13       |   |
+|dbplyr       |1.4.2    |1.4.2      |   |
+|digest       |0.6.22   |0.6.22     |   |
+|dplyr        |0.8.3    |0.8.3      |   |
+|ellipsis     |0.3.0    |0.3.0      |   |
+|evaluate     |0.14     |0.14       |   |
 |fansi        |0.4.0    |0.4.0      |   |
+|farver       |2.0.1    |2.0.1      |   |
 |forcats      |0.4.0    |0.4.0      |   |
 |fs           |1.3.1    |1.3.1      |   |
 |generics     |0.0.2    |0.0.2      |   |
-|ggplot2      |3.1.1    |3.1.1      |   |
+|ggplot2      |3.2.1    |3.2.1      |   |
 |glue         |1.3.1    |1.3.1      |   |
-|haven        |2.1.0    |2.1.0      |   |
+|gtable       |0.3.0    |0.3.0      |   |
+|haven        |2.2.0    |2.2.0      |   |
 |highr        |0.8      |0.8        |   |
-|hms          |0.4.2    |0.4.2      |   |
-|htmltools    |0.3.6    |0.3.6      |   |
-|httr         |1.4.0    |1.4.0      |   |
+|hms          |0.5.2    |0.5.2      |   |
+|htmltools    |0.4.0    |0.4.0      |   |
+|httr         |1.4.1    |1.4.1      |   |
 |jsonlite     |1.6      |1.6        |   |
[…]
