+++ revdep/failures.md
@@ -2,12 +2,12 @@
 
 <details>
 
-* Version: 1.12.0
+* Version: 1.13.5
 * Source code: https://github.com/cran/CNPBayes
 * URL: https://github.com/scristia/CNPBayes
 * BugReports: https://github.com/scristia/CNPBayes/issues
-* Date/Publication: 2018-10-30
-* Number of recursive dependencies: 155
+* Date/Publication: 2019-01-05
+* Number of recursive dependencies: 163
 
 Run `revdep_details(,"CNPBayes")` for more info
 
@@ -15,80 +15,172 @@ Run `revdep_details(,"CNPBayes")` for more info
 
 ## In both
 
+*   checking examples ... ERROR
+    ```
+    ...
+    Running examples in ‘CNPBayes-Ex.R’ failed
+    The error most likely occurred in:
+    
+    > ### Name: ggChains
+    > ### Title: Trace plots of MCMC chains and mixture model densities
+    > ### Aliases: ggChains ggMixture ggMixture,MultiBatchCopyNumber-method
+    > ###   ggMixture,MultiBatchCopyNumberPooled-method
+    > ###   ggMixture,MultiBatchModel-method ggMixture,MultiBatch-method
+    > ###   ggMixture,MultiBatchPooled-method ggChains,MultiBatchModel-method
+    > ###   ggChains,MultiBatchPooled-method
+    > 
+    > ### ** Examples
+    > 
+    >   sb <- SingleBatchModelExample
+    >   iter(sb) <- 1000
+    >   burnin(sb) <- 100
+    >   sb <- posteriorSimulation(sb)
+    >   fig.chains <- ggChains(sb)
+    Error: 1 components of `...` had unexpected names.
+    
+    We detected these problematic arguments:
+    ```
+
 *   R CMD check timed out
     
 
+*   checking whether package ‘CNPBayes’ can be installed ... WARNING
+    ```
+    Found the following significant warnings:
+      Warning: package ‘IRanges’ was built under R version 3.6.1
+      Warning: package ‘S4Vectors’ was built under R version 3.6.1
+      Warning: package ‘GenomicRanges’ was built under R version 3.6.1
+    See ‘/Users/hadley/Documents/tidyverse/tidyverse/revdep/checks.noindex/CNPBayes/new/CNPBayes.Rcheck/00install.out’ for details.
+    ```
+
+*   checking for missing documentation entries ... WARNING
+    ```
+    ...
+      generic 'sigma' and siglist 'MultiBatchPooled'
+      generic 'sigma<-' and siglist 'MixtureModel'
+      generic 'sigma<-' and siglist 'MultiBatchPooled'
+      generic 'tau2' and siglist 'MultiBatch'
+      generic 'theta' and siglist 'MultiBatch'
+      generic 'theta<-' and siglist 'McmcChains,ANY'
+      generic 'theta<-' and siglist 'MixtureModel,ANY'
+      generic 'theta<-' and siglist 'MultiBatch,matrix'
+      generic 'theta<-' and siglist 'MultiBatchModel,ANY'
+      generic 'thin' and siglist 'MultiBatch'
+      generic 'thin' and siglist 'MultiBatchList'
+      generic 'thin<-' and siglist 'McmcParams,numeric'
+      generic 'thin<-' and siglist 'MultiBatch,numeric'
+      generic 'thin<-' and siglist 'MultiBatchList,numeric'
+      generic 'triodata_lrr' and siglist 'TrioBatchModel'
+      generic 'z' and siglist 'MultiBatch'
+      generic 'zFreq' and siglist 'MultiBatch'
[…]
