+++ cran-comments.md
@@ -11,7 +11,7 @@
 
 ## revdepcheck results
 
-We checked 67 reverse dependencies (62 from CRAN + 5 from BioConductor), comparing R CMD check results across CRAN and dev versions of this package.
+We checked 91 reverse dependencies (84 from CRAN + 7 from BioConductor), comparing R CMD check results across CRAN and dev versions of this package.
 
  * We saw 0 new problems
  * We failed to check 4 packages
