### 2019-11-19 #1686 @karthik
@hadley This would be a good time to archive on Zenodo and share the Zenodo DOI here. Since you have so many authors you should make sure that Zenodo lists them in order before finishing that step.
