### 2019-11-19 #1686 @hadley
Is there any specific zenodo advice for JOSS or is it obvious what I should be doing?

One complexity is that the authors of the JOSS paper are the contributing authors across the entire tidyverse, as opposed to the single author (me) of the tidyverse package.
