### 2019-11-19 #1686 @hadley
Or in the interests of simplicity, should I just add the authors of this paper to the tidyverse package?
