### 2019-11-19 #1686 @karthik
Just the authors of this paper.
