### 2019-11-19 fd1db42
Automate tidyverse notification emails
