### 2019-11-19 e69772b
Add vignette metadata
