### 2019-11-19 30bbd84
Eliminate revdepcheck dependency
