### 2019-11-19 #1686 @hadley
Working on it now. While doing so I noticed that the citation you gave me is slightly off — it should be "Welcome to the tidyverse" not "Welcome to the Tidyverse"
