### 2019-11-19 7c33f76
Add citation
