+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: tidyverse
 Title: Easily Install and Load the 'Tidyverse'
-Version: 1.2.1.9000
+Version: 1.3.0
 Authors@R: 
     c(person(given = "Hadley",
              family = "Wickham",
