### 2019-11-19 #1686 @hadley
I have reserved 10.5281/zenodo.3547813, and will upload once I have re-assured myself this will definitely be the same as the version that appears on CRAN.
