+++ cran-comments.md
@@ -1,6 +1,6 @@
 ## Test environments
 
-* local: darwin15.6.0-3.5.1
+* local: darwin15.6.0-3.6.0
 * travis: 3.1, 3.2, 3.3, oldrel, release, devel
 * r-hub: windows-x86_64-devel, ubuntu-gcc-release, fedora-clang-devel
 * win-builder: windows-x86_64-devel
