### 2019-11-20 #1686 @hadley
Ok, published at https://zenodo.org/record/3547813#.XdV2nC3MzUI
