### 2019-11-20 #1686 @karthik
This is ready to publish. Over you to @openjournals/joss-eics
