+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidyverse (development version)
+
 # tidyverse 1.3.0
 
 * The tidyverse now has an associated paper at the 
