### 2019-11-21 #1686 @kthyng
We're just waiting now for the DOI to resolve, should be any minute. In the meantime, congrats to @hadley and the other authors on your new paper!! Thanks to @karthik for editing and for the time and expertise of our reviewers, @ldecicco-USGS and @jeffreyhanson. 🎉 🎉
