### 2019-11-21 #1686 @hadley
Thanks everyone — we greatly appreciate all your hard work making this paper happen!
